Sets up client side autocomplete
Sets up a simple autocomplete that posts the school names to the
serverm, the server responds with an object that matches those school
names with the respoective schools that they refer to, if it is
undefenid than we can enter the school in ourselves.

Client side it pulls from a google sheet full of all the schools in the
world currently.